Moving app dir to more useful name
