+++ poli/models/analysis.py
@@ -1,7 +1,7 @@
 """
     Definition for each analysis result
 """
-from app import db, ma
+from poli import db, ma
 
 
 class AnalysisResult(db.Model):
