+++ poli/models/yara_rule.py
@@ -8,9 +8,9 @@
 '''
 from datetime import datetime
 
-from app import db, ma
-from app.models.models import TLPLevel
-from app.models.analysis import AnalysisResultSchema
+from poli import db, ma
+from poli.models.models import TLPLevel
+from poli.models.analysis import AnalysisResultSchema
 
 
 class YaraRule(db.Model):
