+++ utils/db_migrate.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 import imp
 from migrate.versioning import api
-from app import db
+from poli import db
 from config import SQLALCHEMY_DATABASE_URI
 from config import SQLALCHEMY_MIGRATE_REPO
 v = api.db_version(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO)
